Big refactoring class-core away
