+++ core/class-wc-blocks.php
@@ -1,5 +1,5 @@
 <?php
-namespace Woo_Pakettikauppa_Core;
+namespace Woo_Posti_Core;
 
 if ( ! defined('ABSPATH') ) {
   exit();
@@ -17,11 +17,11 @@ if ( ! class_exists(__NAMESPACE__ . '\Wc_Blocks') ) {
    */
   class Wc_Blocks {
     /**
-     * @var Core
+     * @var \Woo_Posti_Shipping
      */
     private $core = null;
 
-    public function __construct( Core $plugin ) {
+    public function __construct( \Woo_Posti_Shipping $plugin ) {
       $this->core = $plugin;
     }
