+++ core/class-wc-hpos.php
@@ -1,5 +1,5 @@
 <?php
-namespace Woo_Pakettikauppa_Core;
+namespace Woo_Posti_Core;
 
 if ( ! defined('ABSPATH') ) {
   exit();
@@ -19,11 +19,11 @@ if ( ! class_exists(__NAMESPACE__ . '\Wc_Hpos') ) {
    */
   class Wc_Hpos {
     /**
-     * @var Core
+     * @var \Woo_Posti_Shipping
      */
     public $core = null;
 
-    public function __construct( Core $plugin ) {
+    public function __construct( \Woo_Posti_Shipping $plugin ) {
       $this->core = $plugin;
     }
